#5 ferat: refactoring Setting -> Profile
