+++ src/App.tsx
@@ -6,11 +6,11 @@ import Blog from './component/Blog.tsx';
 import Home from './component/Home.tsx';
 import Contact from './component/Contact.tsx';
 import Register from './component/Register.tsx';
-import Setting from './component/Setting.tsx';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import { LoadingProvider, useLoading } from './contexts/LoadingContext.tsx';
 import Loading from './component/Loading.tsx';
+import Profile from './component/Profile.tsx';
 
 const router = createHashRouter([
     {
@@ -35,8 +35,8 @@ const router = createHashRouter([
                 element: <Register />
             },
             {
-                path: '/setting',
-                element: <Setting />
+                path: '/profile',
+                element: <Profile />
             }
         ]
     },
